Rename 'request' to 'requests'
