+++ src/leapfrogai_api/routers/openai/audio.py
@@ -5,7 +5,7 @@ from typing import Annotated
 from fastapi import HTTPException, APIRouter, Depends
 from leapfrogai_api.backend.grpc_client import create_transcription, create_translation
 from leapfrogai_api.backend.helpers import read_chunks
-from leapfrogai_api.typedef.request import (
+from leapfrogai_api.typedef.requests import (
     CreateTranscriptionRequest,
     CreateTranscriptionResponse,
     CreateTranslationRequest,
