+++ tests/integration/api/test_runs.py
@@ -18,7 +18,7 @@ from leapfrogai_api.routers.openai.requests.create_message_request import (
 from leapfrogai_api.routers.openai.requests.create_thread_request import (
     CreateThreadRequest,
 )
-from leapfrogai_api.typedef.request import (
+from leapfrogai_api.typedef.requests import (
     RunCreateParamsRequest,
 )
 from leapfrogai_api.routers.openai.requests.thread_run_create_params_request import (
